requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.28.1
pydantic==2.10.3
PyYAML==6.0.2
numpy==2.1.3
pandas==2.2.3
scipy==1.14.1
matplotlib==3.9.2
psutil==6.1.0
pynvml==12.0.0
